+++ company-tests.el
@@ -237,25 +237,26 @@
   (with-temp-buffer
     (save-window-excursion
     (set-window-buffer nil (current-buffer))
-    (insert "aaaa\n bb\nccccc\nddd")
+    (insert "aaaa\n  bb\nccccccc\nddd")
     (search-backward "bb")
     (let ((col (company--column))
           (company-candidates-length 2)
           (company-candidates '("123" "45")))
       (company-pseudo-tooltip-show (company--row) col 0)
       (let ((ov company-pseudo-tooltip-overlay))
-        (should (eq (overlay-get ov 'company-width) 3))
+        ;; With margins.
+        (should (eq (overlay-get ov 'company-width) 5))
         ;; FIXME: Make it 2?
         (should (eq (overlay-get ov 'company-height) company-tooltip-limit))
         (should (eq (overlay-get ov 'company-column) col))
         (should (string= (overlay-get ov 'company-after)
-                         " 123\nc45 c\nddd\n")))))))
+                         "  123 \nc 45  c\nddd\n")))))))
 
 (ert-deftest company-create-lines-shows-numbers ()
   (let ((company-show-numbers t)
         (company-candidates '("x" "y" "z"))
         (company-candidates-length 3))
-    (should (equal '("x 1" "y 2" "z 3")
+    (should (equal '(" x 1 " " y 2 " " z 3 ")
                    (company--create-lines 0 999)))))
 
 (ert-deftest company-column-with-composition ()